reconfigure auto auth and use IdsClientConfig
