+++ libs/auth/scopes/src/lib/fishingLicense.scope.ts
@@ -1,3 +1,3 @@
 export enum FishingLicenseScope {
-  main = '@fiskistofa/main',
+  fishingLicense = '@fiskistofa.is/fishing-license',
 }
