+++ libs/clients/fishing-licence/src/index.ts
@@ -5,4 +5,4 @@ export {
   FishingLicenceApiClientMock,
 } from './lib/fishing-license-client.mock' //todo remove mock and export from ./gen/fetch
 export * from './lib/types'
-export { FishingLicenseClientConfig } from './lib/fishing-license.config'
+export { FishingLicenseClientConfig } from './lib/FishingLicenseClientConfig'
