+++ libs/clients/fishing-licence/src/lib/FishingLicenseApiFactoryConfig.ts
@@ -1,18 +1,31 @@
 import { createEnhancedFetch } from '@island.is/clients/middlewares'
-import { ConfigType, XRoadConfig } from '@island.is/nest/config'
+import {
+  ConfigType,
+  IdsClientConfig,
+  XRoadConfig,
+} from '@island.is/nest/config'
 import { Configuration } from '../../gen/fetch'
-import { FishingLicenseClientConfig } from './fishing-license.config'
+import { FishingLicenseClientConfig } from './FishingLicenseClientConfig'
 
 export const FishingLicenseApiFactoryConfig = (
   xRoadConfig: ConfigType<typeof XRoadConfig>,
   config: ConfigType<typeof FishingLicenseClientConfig>,
+  idsClientConfig: ConfigType<typeof IdsClientConfig>,
 ) =>
   new Configuration({
     fetchApi: createEnhancedFetch({
       name: 'clients-fishing-license',
       treat400ResponsesAsErrors: true,
       logErrorResponseBody: true,
-      ...config.fetch,
+      autoAuth: idsClientConfig.clientSecret
+        ? {
+            mode: 'tokenExchange',
+            issuer: idsClientConfig.issuer,
+            clientId: idsClientConfig.clientId,
+            clientSecret: idsClientConfig.clientSecret,
+            scope: config.scope,
+          }
+        : undefined,
     }),
     basePath: `${xRoadConfig.xRoadBasePath}/r1/${config.xRoadServicePath}/`,
     headers: {
