+++ libs/clients/fishing-licence/src/lib/UtgerdirApiProvider.ts
@@ -2,11 +2,12 @@ import { Provider } from '@nestjs/common'
 import { UtgerdirApi } from '../../gen/fetch'
 import {
   ConfigType,
+  IdsClientConfig,
   LazyDuringDevScope,
   XRoadConfig,
 } from '@island.is/nest/config'
-import { FishingLicenseClientConfig } from './fishing-license.config'
-import { FishingLicenseApiFactoryConfig } from './fishing-license-api-factory.config'
+import { FishingLicenseClientConfig } from './FishingLicenseClientConfig'
+import { FishingLicenseApiFactoryConfig } from './FishingLicenseApiFactoryConfig'
 
 export const UtgerdirApiProvider: Provider<UtgerdirApi> = {
   provide: UtgerdirApi,
@@ -14,7 +15,11 @@ export const UtgerdirApiProvider: Provider<UtgerdirApi> = {
   useFactory: (
     xRoadConfig: ConfigType<typeof XRoadConfig>,
     config: ConfigType<typeof FishingLicenseClientConfig>,
-  ) => new UtgerdirApi(FishingLicenseApiFactoryConfig(xRoadConfig, config)),
+    idsConfig: ConfigType<typeof IdsClientConfig>,
+  ) =>
+    new UtgerdirApi(
+      FishingLicenseApiFactoryConfig(xRoadConfig, config, idsConfig),
+    ),
 
   inject: [XRoadConfig.KEY, FishingLicenseClientConfig.KEY],
 }
